bots/telegram_bot/requirements.txt
python-telegram-bot[job-queue]
requests
python-dotenv
faster-whisper
pyyaml
